pyproject.toml
[project]
name = "training-llm-application-development"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere==5.19.0",
    "decorator==5.2.1",
    "faiss-cpu==1.12.0",
    "langchain==1.0.1",
    "langchain-chroma==1.0.0",
    "langchain-community==0.4",
    "langchain-core==1.0.0",
    "langchain-experimental==0.0.42",
    "langchain-openai==1.0.0",
    "langchain-tavily>=0.2.12",
    "langchain-text-splitters==1.0.0",
    "langgraph-checkpoint-sqlite==3.0.0",
    "langgraph==1.0.1",
    "nltk==3.9.2",
    "numpy==2.3.4",
    "openai==2.6.0",
    "pydantic-settings==2.11.0",
    "python-dotenv==1.1.1",
    "rank-bm25==0.2.2",
    "rapidfuzz==3.14.1",
    "retry==0.9.2",
    "streamlit-feedback==0.1.4",
    "streamlit==1.50.0",
    "tavily-python==0.7.12",
    "langchain-mcp-adapters>=0.1.11",
    "wandb>=0.23.1",
    "weave>=0.52.22",
    "litellm>=1.80.0",
]

[tool.uv]
dev-dependencies = ["jupyter==1.1.1"]
